Add refactored input event with event code generation
