+++ services/install_reqs.sh
@@ -9,4 +9,5 @@ pip3 install events_slack_app/slack_response/
 pip3 install events_slack_app/slack_command/
 pip3 install poller/daily_poller/
 pip3 install structured_mysql/batch_job_aurora/
+pip3 install structured_mysql/update/
 pip3 install integration_tests/util/
